pyproject.toml
[project]
name = "imkerutils"
version = "0.1.0"
description = "Image processing utilities, especially for working alongside LLM/Diff/VIT models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.21.0",
    "Pillow>=10",
    "pytest>=9.0.2",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

# Package discovery: include imkerutils and all subpackages (imkerutils.*)
[tool.setuptools.packages.find]
where = ["."]
include = ["imkerutils*"]

[project.scripts]
extract-tile = "imkerutils.tiling_utils.extract_tile:main"
paste-tile = "imkerutils.tiling_utils.paste_tile:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: tests that hit real external services (network/billing); opt-in only",
]
